+++ packages/flutter_user/lib/src/screens/email_password_login_form.dart
@@ -5,6 +5,7 @@ import "package:flutter_accessibility/flutter_accessibility.dart";
 import "package:flutter_user/src/models/login/login_options.dart";
 import "package:flutter_user/src/services/local_auth.dart";
 import "package:flutter_user/src/widgets/biometrics_button.dart";
+import "package:flutter_user/src/widgets/optional_spacer.dart";
 
 class EmailPasswordLoginForm extends StatefulWidget {
   /// Constructs an [EmailPasswordLoginForm] widget.
@@ -263,9 +264,9 @@ class _EmailPasswordLoginFormState extends State<EmailPasswordLoginForm> {
                               ),
                             ),
                             forgotPasswordButton,
-                            if (options.spacers.spacerAfterForm != null) ...[
-                              Spacer(flex: options.spacers.spacerAfterForm!),
-                            ],
+                            ...buildOptionalSpacer(
+                              options.spacers.spacerAfterForm,
+                            ),
                             if (options
                                 .biometricsOptions.loginWithBiometrics) ...[
                               Row(
@@ -290,9 +291,9 @@ class _EmailPasswordLoginFormState extends State<EmailPasswordLoginForm> {
                             if (widget.onRegister != null) ...[
                               registerButton,
                             ],
-                            if (options.spacers.spacerAfterButton != null) ...[
-                              Spacer(flex: options.spacers.spacerAfterButton!),
-                            ],
+                            ...buildOptionalSpacer(
+                              options.spacers.spacerAfterButton,
+                            ),
                           ],
                         ),
                       ),
@@ -324,9 +325,9 @@ class _LoginTitle extends StatelessWidget {
     var theme = Theme.of(context);
     return Column(
       children: [
-        if (options.spacers.spacerBeforeTitle != null) ...[
-          Spacer(flex: options.spacers.spacerBeforeTitle!),
-        ],
+        ...buildOptionalSpacer(
+          options.spacers.spacerBeforeTitle,
+        ),
         if (title != null) ...[
           Align(
             alignment: Alignment.topCenter,
@@ -336,9 +337,9 @@ class _LoginTitle extends StatelessWidget {
             ),
           ),
         ],
-        if (options.spacers.spacerAfterTitle != null) ...[
-          Spacer(flex: options.spacers.spacerAfterTitle!),
-        ],
+        ...buildOptionalSpacer(
+          options.spacers.spacerAfterTitle,
+        ),
         if (subtitle != null) ...[
           Align(
             alignment: Alignment.topCenter,
@@ -348,18 +349,18 @@ class _LoginTitle extends StatelessWidget {
             ),
           ),
         ],
-        if (options.spacers.spacerAfterSubtitle != null) ...[
-          Spacer(flex: options.spacers.spacerAfterSubtitle!),
-        ],
+        ...buildOptionalSpacer(
+          options.spacers.spacerAfterSubtitle,
+        ),
         if (options.image != null) ...[
           Padding(
             padding: const EdgeInsets.all(16),
             child: options.image,
           ),
         ],
-        if (options.spacers.spacerAfterImage != null) ...[
-          Spacer(flex: options.spacers.spacerAfterImage!),
-        ],
+        ...buildOptionalSpacer(
+          options.spacers.spacerAfterImage,
+        ),
       ],
     );
   }