+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Unreleased
-
+- Removed the default values for colors in the LoginOptions, RegistrationOptions, and ForgotPasswordOptions.
 - Added rest_user_repository package which is a REST implementation of the UserRepositoryInterface.
 - Changed afterLoginScreen to a nullable Widget so a screen isn't automatically pushed after login.
 - Moved the RegistrationOptions and ForgotPasswordOptions to the FlutterUserOptions.